pyproject.toml
[project]
name = "gemini-research-mcp"
version = "0.1.0"
description = "Unified Gemini research partner — video analysis, deep research, content extraction via MCP"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Fausto" }]
dependencies = [
    "fastmcp>=2.0",
    "google-genai>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
]

[project.scripts]
gemini-research-mcp = "gemini_research_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_research_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
